CoDA/requirements.txt
aiohttp==3.10.11
jaraco.collections==5.1.0
nltk==3.8.1
openai==2.2.0
pandas==1.5.3
pip-chill==1.0.3
platformdirs==4.2.2
pytorch-lightning==1.9.5
scikit-learn==1.3.2
sentencepiece==0.1.99
textstat==0.7.4
tomli==2.0.1
torchaudio==0.12.1+cu116
torchvision==0.13.1+cu116
transformers==4.14.1